Add: flag to disable cron job

+++ prices-crawler-content-service/src/main/java/io/github/pricescrawler/content/service/scheduled/BackgroundService.java
@@ -4,6 +4,7 @@ import io.github.pricescrawler.content.service.product.cache.ProductCacheService
 import io.github.pricescrawler.content.service.product.list.ProductListService;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +14,9 @@ public class BackgroundService {
     private final ProductListService productListService;
     private final ProductCacheService localProductCacheService;
     private final ProductCacheService mongoDbProductCacheService;
+
+    @Value("${prices.crawler.background.service.cron.enabled:false}")
+    private boolean enabled;
 
     public BackgroundService(ProductListService productListService,
                              @Qualifier("inMemoryProductCacheService")
@@ -26,9 +30,11 @@ public class BackgroundService {
 
     @Scheduled(cron = "${prices.crawler.background.service.cron:0 0 0 * * *}")
     public void startBackgroundService() {
-        log.info("Starting Background Service");
-        productListService.deleteOutdatedProductLists();
-        localProductCacheService.deleteOutdatedProductSearchResults();
-        mongoDbProductCacheService.deleteOutdatedProductSearchResults();
+        if (enabled) {
+            log.info("Starting Background Service");
+            productListService.deleteOutdatedProductLists();
+            localProductCacheService.deleteOutdatedProductSearchResults();
+            mongoDbProductCacheService.deleteOutdatedProductSearchResults();
+        }
     }
 }
